requirements.txt
numpy>=1.26.4
pandas>=2.3.1
POT>=0.9.5
pykan>=0.2.8
scanpy>=1.10.4
scikit-learn>=1.7.1
scipy>=1.11.4
torch==2.1.0+cu121
tqdm>=4.67.1
